seminar_9/requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.24
numpy==1.23.3
pandas==1.5.0
python-dateutil==2.8.2
python-telegram-bot==13.14
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2022.4
tzlocal==4.2
isOdd==0.1.2
